+++ docs/RELEASE_NOTES.md
@@ -1,3 +1,15 @@
+# TerminusDB Server v11.2-rc4 Release Notes
+
+## Enhancements
+* Vanilla JSON support with deduplication is back, with sys:JSON type!
+* Top level sys:JSON values are now possible, except for null
+* sys:JSON numbers are processed as xsd:decimal precision (beyond float/double)
+* sys:JSON handles arbitrary precision JSON accurately up to a max of 2048 bytes of storage
+
+## Other
+* 5x faster Rust serde JSON parsing, instead of Prolog built in JSON parser
+  * From ~80,000 parses/sec to ~402,000 parses/sec on a 2023 M3 Max MacBook Pro
+
 # TerminusDB Server v11.2-rc3 Release Notes
 
 Release Candidate 3 brings improved error handling and observability, makes it possible again to run the dashboard component (or other static webapp dashboards), and further strengthens the core of TerminusDB with additional bug fixes.